Change HTTPClient protocol to return BothamAPIClientError
